Include timezone as a schedule argument (#203)
Timezone is a pretty important part of any schedule - moving it to the function argument should make it clear how to do so.

I saw the kwargs, so I tried `timeZone` (what the API wants) and `time_zone` (what the python scheduler library wants). Had to go through the source code to find out it was `timezone`, so I hope this change makes things easier for other newcomers!